migrate tests to use new NZSHM22 package namespace (junit + gradle all OK)
