--- test/resources/NSHM_NZ/inversion/patch_4_10.csv
@@ -1,10 +1,0 @@
-along_strike_index,down_dip_index,lon1(deg),lat1(deg),lon2(deg),lat2(deg),dip (deg),top_depth (km),bottom_depth (km),slip_deficit (mm/yr),tile_geometry
-3,9,172.05718990191556,-43.02716092186062,171.94629898533478,-43.06580050196082,12.05019252859843,36.59042136801586,38.67810629370413,0.0,"[(4, 9), (3, 10), (4, 10)]"
-3,10,171.96728185661328,-42.9683866660248,171.8561649665836,-43.00650585381854,18.043805849579094,38.292799925819594,41.39024032826338,0.0,"[(4, 11), (4, 10), (4, 9), (3, 9), (3, 11)]"
-3,11,171.88236259240452,-42.90546756100804,171.77605503938062,-42.950241141732384,26.97629350005077,41.24616883544199,45.78238684712409,0.0,"[(4, 11), (4, 10), (3, 10)]"
-4,9,172.1486674175689,-42.961161741635415,172.04164019685234,-43.005133654155124,14.222450383168196,37.01693572221088,39.473808000399664,0.0,"[(5, 10), (5, 9), (3, 9), (3, 10), (4, 10)]"
-4,10,172.05856409919377,-42.90499137556964,171.9492806375649,-42.94571229036213,19.42590467503754,39.19738172958217,42.52325722471111,0.0,"[(5, 11), (5, 10), (5, 9), (4, 9), (3, 9), (3, 10), (3, 11), (4, 11)]"
-4,11,171.97306892041814,-42.844263450308,171.86694649374164,-42.889143194547096,27.05696391501034,42.221710291867204,46.7704715010719,0.0,"[(5, 11), (5, 10), (4, 10), (3, 10), (3, 11)]"
-5,9,172.24243850307948,-42.89748270300359,172.13921345652957,-42.9459933172786,16.27630017436574,37.35528596441705,40.15798267512191,0.0,"[(4, 9), (4, 10), (5, 10)]"
-5,10,172.15279087916238,-42.84325280144621,172.0457520503184,-42.88694084710267,20.6448807647893,39.88466001719286,43.41040773562808,0.0,"[(5, 9), (4, 9), (4, 10), (4, 11), (5, 11)]"
-5,11,172.0673892180764,-42.78509363274145,171.96182184685478,-42.83055360508922,26.764579651521238,43.18726230205054,47.69051886699764,0.0,"[(5, 10), (4, 10), (4, 11)]"
